NutsKeyResolver to separate file
