+++ vdr/didservice/nuts_resolvers.go
@@ -1,0 +1,46 @@
+/*
+ * Copyright (C) 2023 Nuts community
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package didservice
+
+import (
+	"crypto"
+	"github.com/nuts-foundation/nuts-node/crypto/hash"
+	"github.com/nuts-foundation/nuts-node/vdr/types"
+)
+
+// NutsKeyResolver implements the NutsKeyResolver interface.
+type NutsKeyResolver struct {
+	Resolver types.DocResolver
+}
+
+func (r NutsKeyResolver) ResolvePublicKey(kid string, sourceTransactionsRefs []hash.SHA256Hash) (crypto.PublicKey, error) {
+	// try all keys, continue when err == types.ErrNotFound
+	for _, h := range sourceTransactionsRefs {
+		publicKey, err := resolvePublicKey(r.Resolver, kid, types.ResolveMetadata{
+			SourceTransaction: &h,
+		})
+		if err == nil {
+			return publicKey, nil
+		}
+		if err != types.ErrNotFound {
+			return nil, err
+		}
+	}
+
+	return nil, types.ErrNotFound
+}
